Add inline plot support for jupyter notebooks
Calling auto.plot() from a jupyter notebook previously used the
TkAgg backend and opened a new TkInter window. This interfered with
the jupyter eventloop and made Auto difficult to use from
jupyter. This patch avoids the problem, by not interfering with the
matplotlib backend, if we run from within a notebook.  This results
in (non-interactive) inline plots inside the browser window, similar
to matplotlib figures after "%matplotlib inline". An example is given
in "demos/cusp/cusp.ipynb"